manual map layout can be saved and restored

+++ src/java/netxms-base/src/main/java/org/netxms/base/NXCPVariable.java
@@ -243,7 +243,7 @@ public class NXCPVariable
 			switch(variableType)
 			{
 				case TYPE_INTEGER:
-					integerValue = (long)in.readUnsignedInt();
+					integerValue = in.readUnsignedInt();
 					realValue = integerValue.doubleValue();
 					stringValue = integerValue.toString();
 					break;
